// miniapp/src/types/openapi.ts
export type RegionItem = {
  region_id: number;
  parent_id: number;
  level: number;
  region_name: string;
  longitude: string;
  latitude: string;
  pinyin: string;
  az_no: string;
};